refactor: 테마 스위치 로직 훅으로 분리
